+++ app/domain/booking_states/overdue.rb
@@ -22,7 +22,7 @@ module BookingStates
     end
 
     infer_transition(to: :upcoming) do |booking|
-      booking.contracts.signed.any? && Invoices::Deposit.of(booking).kept.all?(&:paid)
+      booking.contracts.signed.any? && Invoices::Deposit.of(booking).kept.all?(&:paid?)
     end
 
     def relevant_time
@@ -32,7 +32,7 @@ module BookingStates
     protected
 
     def deposits_paid_checklist_item
-      ChecklistItem.new(:deposit_paid, Invoices::Deposit.of(booking).kept.all?(&:paid),
+      ChecklistItem.new(:deposit_paid, Invoices::Deposit.of(booking).kept.all?(&:paid?),
                         manage_booking_invoices_path(booking, org: booking.organisation.slug))
     end