+++ app/models/invoice.rb
@@ -6,9 +6,9 @@
 #
 #  id                 :bigint           not null, primary key
 #  amount             :decimal(, )      default(0.0)
+#  amount_open        :decimal(, )
 #  discarded_at       :datetime
 #  issued_at          :datetime
-#  paid               :boolean          default(FALSE)
 #  payable_until      :datetime
 #  payment_info_type  :string
 #  print_payment_slip :boolean          default(FALSE)
@@ -41,17 +41,18 @@ class Invoice < ApplicationRecord
   has_one_attached :pdf
   has_one :organisation, through: :booking
 
-  scope :ordered, -> { order(payable_until: :ASC, created_at: :ASC) }
-  scope :unpaid,  -> { kept.where(paid: false) }
-  scope :paid,    -> { kept.where(paid: true) }
-  scope :sent,    -> { kept.where.not(sent_at: nil) }
-  scope :unsent,  -> { kept.where(sent_at: nil) }
-  scope :overdue, ->(at = Time.zone.today) { kept.where(arel_table[:payable_until].lteq(at)) }
-  scope :of, ->(booking) { where(booking: booking) }
+  scope :ordered,  -> { order(payable_until: :ASC, created_at: :ASC) }
+  scope :unpaid,   -> { kept.where(arel_table[:amount_open].gt(0)) }
+  scope :overpaid, -> { kept.where(arel_table[:amount_open].lt(0)) }
+  scope :paid,     -> { kept.where(arel_table[:amount_open].lteq(0)) }
+  scope :sent,     -> { kept.where.not(sent_at: nil) }
+  scope :unsent,   -> { kept.where(sent_at: nil) }
+  scope :overdue,  ->(at = Time.zone.today) { kept.where(arel_table[:payable_until].lteq(at)) }
+  scope :of,       ->(booking) { where(booking: booking) }
   scope :with_default_includes, -> { includes(%i[invoice_parts payments organisation]) }
 
   accepts_nested_attributes_for :invoice_parts, reject_if: :all_blank, allow_destroy: true
-  before_save :set_amount, :set_paid
+  before_save :recalculate
   before_update :generate_pdf, if: :generate_pdf?
   after_create { generate_ref? && generate_ref && save }
   delegate :invoice_address_lines, to: :booking
@@ -76,17 +77,25 @@ class Invoice < ApplicationRecord
     self.ref = invoice_ref_strategy.generate(self)
   end
 
-  def set_amount
-    self.amount = invoice_parts.sum(&:amount)
-  end
-
   def amount_in_cents
     (amount * 100).to_i
   end
 
+  def paid?
+    overpaid? || amount_open.zero?
+  end
+
+  def overpaid?
+    amount_open.negative?
+  end
+
+  def recalculate
+    self.amount = invoice_parts.ordered.inject(0) { |sum, invoice_part| invoice_part.to_sum(sum) }
+    self.amount_open = amount - amount_paid
+  end
+
   def recalculate!
-    set_amount
-    set_paid
+    recalculate
     save
   end
 
@@ -94,16 +103,8 @@ class Invoice < ApplicationRecord
     "#{self.class.model_name.human} #{booking.ref}_#{id}.pdf"
   end
 
-  def amount_open
-    amount - amount_paid
-  end
-
   def amount_paid
     payments.sum(&:amount)
-  end
-
-  def set_paid
-    self.paid = amount_open.zero?
   end
 
   def sent!