+++ app/models/invoice_parts.rb
@@ -1,5 +1,5 @@
 # frozen_string_literal: true
 
 module InvoiceParts
-  TYPES = [InvoiceParts::Add].freeze
+  TYPES = [InvoiceParts::Add, InvoiceParts::Percentage].freeze
 end