+++ app/models/invoice_parts/add.rb
@@ -28,8 +28,5 @@
 
 module InvoiceParts
   class Add < InvoicePart
-    def inject_self(result)
-      result + amount
-    end
   end
 end