+++ app/models/invoice_parts/percentage.rb
@@ -1,0 +1,35 @@
+# frozen_string_literal: true
+
+# == Schema Information
+#
+# Table name: invoice_parts
+#
+#  id         :bigint           not null, primary key
+#  amount     :decimal(, )
+#  breakdown  :string
+#  label      :string
+#  ordinal    :integer
+#  type       :string
+#  created_at :datetime         not null
+#  updated_at :datetime         not null
+#  invoice_id :bigint
+#  usage_id   :bigint
+#
+# Indexes
+#
+#  index_invoice_parts_on_invoice_id  (invoice_id)
+#  index_invoice_parts_on_usage_id    (usage_id)
+#
+# Foreign Keys
+#
+#  fk_rails_...  (invoice_id => invoices.id)
+#  fk_rails_...  (usage_id => usages.id)
+#
+
+module InvoiceParts
+  class Percentage < InvoicePart
+    def calculated_amount
+      sum_of_predecessors * amount / 100
+    end
+  end
+end