+++ app/webpack/models/Home.ts
@@ -1,0 +1,4 @@
+export type Home = {
+  id: number;
+  name: string;
+};