+++ app/webpack/services/csrf.ts
@@ -1,0 +1,6 @@
+export function getCsrfToken(): string {
+  return (
+    (document.querySelector("meta[name=csrf-token]") as HTMLMetaElement)
+      ?.content || ""
+  );
+}